refactor chest dialog functions into actions
