+++ src/features/dialog-manager/actions/close-chest-dialog.js
@@ -1,0 +1,14 @@
+
+export default function closeChestDialog() {
+  return dispatch => {
+
+    dispatch({
+      type: 'SET_CHEST_DATA',
+      payload: false
+    });
+    dispatch({
+      type: 'PAUSE',
+      payload: { pause: false }
+    });
+  }
+}
